fix broken imports after lmdbscan was moved
